+++ cookbooks/adobe_indesign_cs6/metadata.rb
@@ -1,6 +1,7 @@
+name			 "adobe_indesign_cs6"
 maintainer       "Wesleyan University"
-maintainer_email "rchristensen@wesleyan.edu"
+maintainer_email "software@wesleyan.edu"
 license          "All rights reserved"
 description      "Installs Adobe InDesign CS6"
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          "0.2.0"
+version          "0.3.0"